testing for domain package updated (100% coverage)
